chore: moved suggestion menu styles into styles folder
