+++ packages/formula-editor/src/styles/editor.ts
